pyproject.toml
[tool.poetry]
name = "excelschema"
version = "0.1.2.4"
description = "Excel records' parser and schema viewing and validating tools."
authors = ["patarapolw <patarapolw@gmail.com>"]
readme = "README.md"
repository = "https://github.com/patarapolw/excelschema"
homepage = "https://github.com/patarapolw/excelschema"
keywords = ["excel"]

[tool.poetry.dependencies]
python = "*"
python-dateutil = "^2.7"

[tool.poetry.dev-dependencies]
